Fix step 1 onboarding screens by rearranging Vuex store
The Vuex store is not implemented very well right now. This is
the first commit which starts moving things around.

My current idea is that the store should contain most of the data
coming from the backend, and that much of that can be retrieved
during either login or startup. Then we could have a separate
event which periodically fetches updated data, which will then
store that data and the view will be automatically updated. This
way the individual screens do not, for the most part, need to
be fetching data.