Remove all services and make the API call directly
The dual maintenance of having services as well as needing to
pass parameters to axios is not worth it. The caller needs to
know the API calls in detail anyway, simply to pass the correct
data.

This commit also include an example of how to share a sequence
of calls in ~/assets/utils.js. A simple pattern like this can
be used when needed.