Pass data to Stats, not define in Stats
A previous commit changed to define the data for the stats within the
Stats component itself. This creates unnecessary coupling between the
Stats component and anything which needs to use it. The purpose was
to reduce copied code, but the result is worse.

This commit refactors to return to passing the data into the Stats
component. I then created another component to define the data to
pass into the Stats so that these higher level components can be
used, which was the purpose of the previous refactoring.