Use the stored rank to get signatures rather than the live one
We were previously passing the most up-to-date rank to the signatures,
but this can be out-of-date with the version in the server (specifically
in the case of test users, which is where this issue came from). With this
fix, we use the version from the server.

However, the user might have gone up a rank, and so now we won't
offer new signatures. However the correct way of doing this would be
for the server to update the information in the userStats, rather than
the client doing it (in that case the two pieces of data would not be
out of sync). So here we choose consistency rather than being the most
up-to-date.

Closes #126